requirements-dev.txt
aspy.yaml==1.1.2
atomicwrites==1.3.0
attrs==18.2.0
cached-property==1.5.1
certifi==2018.11.29
cfgv==1.4.0
chardet==3.0.4
cracklib==2.9.3
dnspython==1.16.0
identify==1.2.2
idna==2.8
importlib-metadata==0.8
importlib-resources==1.0.2
Jinja2==2.10
jsonschema==2.6.0
ldap3==2.5.2
MarkupSafe==1.1.0
more-itertools==6.0.0
nodeenv==1.3.3
ocflib==2019.2.20.14.19
pathlib2==2.3.3
pexpect==4.6.0
pluggy==0.8.1
ply==3.11
pre-commit==1.14.4
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.5
pycryptodome==3.7.3
pycryptodomex==3.7.3
pyfakefs==3.5.7
PyMySQL==0.9.3
pysmi==0.3.3
pysnmp==4.4.9
pytest==4.3.0
python-dateutil==2.8.0
PyYAML==3.13
redis==3.2.0
requests==2.21.0
requirements-tools==1.2.1
setuptools==40.8.0
six==1.12.0
SQLAlchemy==1.2.18
toml==0.10.0
urllib3==1.24.1
virtualenv==16.4.0
zipp==0.3.3
